caa/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
python-multipart==0.0.6
openpyxl==3.1.2
pydantic==2.5.3
python-dotenv==1.0.0
websockets==12.0